Extract Wrappers from components
