+++ src/components/Statistic.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import styled from 'styled-components'
-import pending from '../images/pending.png'
-import schedule from '../images/schedule.png'
-import decline from '../images/decline.png'
+import pending from '../asserts/images/pending.png'
+import schedule from '../asserts/images/schedule.png'
+import decline from '../asserts/images/decline.png'
+import Wrapper from '../asserts/wrappers/Statistic'
 
 const Statistic = () => {
     return (
@@ -14,7 +14,6 @@ const Statistic = () => {
                         <h2>100</h2>
                     </div>
                     <img src={pending} alt="pending" />
-
                 </div>
                 <div className="box box2">
                     <div>
@@ -22,7 +21,6 @@ const Statistic = () => {
                         <h2>100</h2>
                     </div>
                     <img src={schedule} alt="pending" />
-
                 </div>
                 <div className="box box3">
                     <div>
@@ -30,41 +28,12 @@ const Statistic = () => {
                         <h2>100</h2>
                     </div>
                     <img src={decline} alt="pending" />
-
                 </div>
             </div>
         </Wrapper>
     );
 }
 
-const Wrapper = styled.section`
-    padding-top: 20px;
-    .box_wrapper{
-        display: grid;
-        grid-template-columns: 1fr 1fr 1fr;
-        column-gap: 30px;
-    }
-    .box{
-        display: grid;
-        grid-template-columns: 2fr 1fr;
-        border-radius: 15px;
-        box-shadow: rgba(153, 153, 241, 0.3) 0px 7px 29px 0px;
-        padding: 5px 20px 5px 20px;
-        border: 1px solid rgba(153, 153, 241, 0.3);
-    }
-    .box1{
-        background-color: #C2DED1;
-    }
-    .box2{
-        background-color: #D9D7F1;
-    }
-    .box3{
-        background-color: #8CC0DE;
-    }
-    img{
-        width: 180px;
-    }
 
-`
 
 export default Statistic
